chore(deps): update dependency baseline
- calibrate policy warnings

- upgrade actionable packages

- defer incompatible majors
