+++ apps/mobile/package.json
@@ -29,16 +29,16 @@
     "@fidy/schemas": "workspace:*",
     "@fidy/types": "workspace:*",
     "@fidy/utils": "workspace:*",
-    "@react-native-community/datetimepicker": "8.6.0",
-    "@react-native-community/netinfo": "11.5.2",
-    "@react-navigation/bottom-tabs": "^7.4.0",
-    "@react-navigation/elements": "^2.6.3",
-    "@react-navigation/native": "^7.1.8",
-    "@sentry/cli": "^3.3.3",
+    "@react-native-community/datetimepicker": "9.1.0",
+    "@react-native-community/netinfo": "12.0.1",
+    "@react-navigation/bottom-tabs": "7.15.10",
+    "@react-navigation/elements": "2.9.15",
+    "@react-navigation/native": "7.2.2",
+    "@sentry/cli": "3.4.0",
     "@sentry/react-native": "~7.11.0",
-    "@shopify/flash-list": "2.0.2",
-    "@supabase/supabase-js": "^2.98.0",
-    "@tanstack/react-query": "5.90.5",
+    "@shopify/flash-list": "2.3.1",
+    "@supabase/supabase-js": "2.104.1",
+    "@tanstack/react-query": "5.100.1",
     "burnt": "^0.13.0",
     "date-fns": "^4.1.0",
     "drizzle-orm": "0.45.2",
@@ -68,20 +68,20 @@
     "expo-updates": "~55.0.20",
     "expo-web-browser": "~55.0.14",
     "i18n-js": "^4.5.3",
-    "lucide-react-native": "^0.575.0",
+    "lucide-react-native": "1.11.0",
     "nativewind": "^4",
-    "posthog-react-native": "^4.37.6",
+    "posthog-react-native": "4.43.5",
     "react": "19.2.0",
     "react-dom": "19.2.0",
-    "react-native": "0.83.4",
+    "react-native": "0.83.6",
     "react-native-css-interop": "^0.2.3",
-    "react-native-gesture-handler": "~2.30.0",
+    "react-native-gesture-handler": "2.31.1",
     "react-native-reanimated": "4.2.1",
-    "react-native-safe-area-context": "~5.6.2",
-    "react-native-screens": "~4.23.0",
+    "react-native-safe-area-context": "5.7.0",
+    "react-native-screens": "4.24.0",
     "react-native-svg": "15.15.3",
     "react-native-web": "~0.21.0",
-    "react-native-worklets": "0.7.2",
+    "react-native-worklets": "0.7.4",
     "tailwindcss": "3.4.19",
     "zod": "^4.3.6",
     "zustand": "^5.0.11"
@@ -94,7 +94,7 @@
     "@vitest/coverage-v8": "^4.0.18",
     "babel-plugin-inline-import": "3.0.0",
     "babel-preset-expo": "~55.0.8",
-    "better-sqlite3": "12.8.0",
+    "better-sqlite3": "12.9.0",
     "drizzle-kit": "0.31.9",
     "eslint": "^9.25.0",
     "eslint-config-expo": "~55.0.0",
