+++ dependency-policy.json
@@ -1,0 +1,34 @@
+{
+  "deferredUpdates": [
+    {
+      "name": "@sentry/react-native",
+      "until": "2026-06-01",
+      "reason": "Major v8 migration needs a dedicated Sentry PR with config review, source map verification, and native QA."
+    },
+    {
+      "name": "eslint",
+      "until": "2026-06-01",
+      "reason": "ESLint 10 should wait for eslint-config-expo compatibility confirmation."
+    },
+    {
+      "name": "react-native",
+      "until": "2026-06-01",
+      "reason": "Expo SDK 55 expects react-native 0.83.6; npm latest is outside the supported Expo native set."
+    },
+    {
+      "name": "react-native-reanimated",
+      "until": "2026-06-01",
+      "reason": "react-native-reanimated 4.3.0 requires react-native-worklets 0.8.x, while Expo SDK 55 expects worklets 0.7.4."
+    },
+    {
+      "name": "react-native-worklets",
+      "until": "2026-06-01",
+      "reason": "Expo SDK 55 expects react-native-worklets 0.7.4; npm latest is outside the supported Expo native set."
+    },
+    {
+      "name": "tailwindcss",
+      "until": "2026-06-01",
+      "reason": "Tailwind 4 requires NativeWind compatibility proof and a dedicated styling migration."
+    }
+  ]
+}
