+++ package.json
@@ -27,7 +27,7 @@
     "lint:complexity": "bun ./scripts/run-lizard-strict.ts",
     "lint:file-size": "bun ./scripts/check-max-file-lines.ts --enforce",
     "lint:theme-colors": "bun ./scripts/check-theme-color-literals.ts --enforce",
-    "lint:dependency-policy": "bun ./scripts/check-dependency-policy.ts --enforce --stale-days 30",
+    "lint:dependency-policy": "bun ./scripts/check-dependency-policy.ts --enforce --stale-days 365",
     "lint:brands": "bun scripts/check-branded-boundaries.ts",
     "lint:feature-public-imports": "bun ./scripts/check-feature-public-imports.ts --enforce",
     "lint:architecture": "bun run lint:feature-public-imports && bunx dependency-cruiser --config .dependency-cruiser.cjs apps/mobile packages",
@@ -46,7 +46,7 @@
     "test": "bun run test:brands && bun run test:bootstrap-registry && bun run test:feature-public-imports && bun run test:file-size && bun run test:theme-colors && bun run test:complexity && bun run test:mobile",
     "test:coverage": "bun run --cwd apps/mobile --shell=bun test:coverage",
     "perf:bundle:mobile": "bun run --cwd apps/mobile --shell=bun bundle:atlas",
-    "lint": "bunx biome check .",
+    "lint": "bunx oxlint .",
     "format": "bunx biome format --write .",
     "verify": "bun run lint && bun run lint:brands && bun run lint:mobile && bun run lint:complexity && bun run lint:file-size && bun run lint:theme-colors && bun run lint:dependency-policy && bun run lint:architecture && bun run typecheck && bun run test"
   },
@@ -55,8 +55,9 @@
     "bun-types": "1.3.10",
     "dependency-cruiser": "17.3.10",
     "jscpd": "4.0.9",
-    "knip": "6.3.1",
+    "knip": "6.6.3",
     "lefthook": "2.1.1",
+    "oxlint": "1.61.0",
     "typescript": "6.0.3",
     "xcodebuildmcp": "2.3.2"
   }
